Trying to refresh my fork with the original. If Runestone sees this I did it wrong…oops.

+++ controllers/default.py
@@ -2,12 +2,12 @@
 import json
 import os
 import requests
-from urllib import unquote
-from urllib2 import HTTPError
+from six.moves.urllib.parse import unquote
+from six.moves.urllib.error import HTTPError
 import logging
 
 from gluon.restricted import RestrictedError
-from gluon.tools import addrow
+from stripe_form import StripeForm
 
 logger = logging.getLogger(settings.logger)
 logger.setLevel(settings.log_level)
@@ -16,7 +16,6 @@ logger.setLevel(settings.log_level)
 def user():
     # this is kinda hacky but it's the only way I can figure out how to pre-populate
     # the course_id field
-
     if 'everyday' in request.env.http_host:
         redirect('http://interactivepython.org/runestone/everyday')
 
@@ -83,28 +82,8 @@ def user():
             form.record.update_record(**dict(form.vars))
             # auth.user session object doesn't automatically update when the DB gets updated
             auth.user.update(form.vars)
+            # TODO: Why is this necessary?
             auth.user.course_name = db(db.auth_user.id == auth.user.id).select()[0].course_name
-            # problem is that
-            in_db = db(
-                (db.user_courses.user_id == auth.user.id) & (db.user_courses.course_id == auth.user.course_id)).select()
-            db_check = []
-            for row in in_db:
-                db_check.append(row)
-            if not db_check:
-                db.user_courses.insert(user_id=auth.user.id, course_id=auth.user.course_id)
-            res = db(db.chapters.course_id == auth.user.course_name)
-            logger.debug("PROFILE checking for progress table %s ", res)
-            if res.count() > 0:
-                chapter_label = res.select().first().chapter_label
-                if db((db.user_sub_chapter_progress.user_id == auth.user.id) &
-                              (db.user_sub_chapter_progress.chapter_id == chapter_label)).count() == 0:
-                    db.executesql('''
-                       INSERT INTO user_sub_chapter_progress(user_id, chapter_id,sub_chapter_id, status)
-                       SELECT %s, chapters.chapter_label, sub_chapters.sub_chapter_label, -1
-                       FROM chapters, sub_chapters where sub_chapters.chapter_id = chapters.id and chapters.course_id = %s;
-                    ''', (auth.user.id, auth.user.course_name))
-            else:
-                session.flash = 'This course is not set up for tracking progress'
             # Add user to default section for course.
             sect = db((db.sections.course_id == auth.user.course_id) &
                       (db.sections.name == form.vars.section)).select(db.sections.id).first()
@@ -116,10 +95,20 @@ def user():
             # add a row to section_users for this user with the section selected.
             redirect(URL('default', 'index'))
 
+    if 'register' in request.args(0):
+        # The validation function ``IS_COURSE_ID`` in ``models/db.py`` changes the course name supplied to a course ID. If the overall form doesn't validate, the value when the form is re-displayed with errors will contain the ID instead of the course name. Change it back to the course name. Note: if the user enters a course for the course name, it will be displayed as the corresponding course name after a failed validation. I don't think this case is important enough to fix.
+        try:
+            course_id = int(form.vars.course_id)
+        except:
+            pass
+        else:
+            # Look up the course name based on the ID.
+            form.vars.course_id = getCourseNameFromId(course_id)
+
     # this looks horrible but it seems to be the only way to add a CSS class to the submit button
     try:
         form.element(_id='submit_record__row')[1][0]['_class'] = 'btn btn-default'
-    except AttributeError:  # not all auth methods actually have a submit button (e.g. user/not_authorized)
+    except (AttributeError, TypeError):  # not all auth methods actually have a submit button (e.g. user/not_authorized)
         pass
     return dict(form=form)
 
@@ -132,6 +121,48 @@ def download(): return response.download(request, db)
 
 
 def call(): return service()
+
+
+# Determine the student price for a given ``course_id``. Returns a value in cents, where 0 cents indicates a free book.
+def _course_price(course_id):
+    # Look for a student price for this course.
+    course = db(db.courses.id == course_id).select(db.courses.student_price, db.courses.base_course).first()
+    assert course
+    price = course.student_price
+    # Only look deeper if a price isn't set (even a price of 0).
+    if price is None:
+        # See if the base course has a student price.
+        base_course = db(db.courses.course_name == course.base_course).select(db.courses.student_price).first()
+        # If this is already a base course, we're done.
+        if base_course:
+            price = base_course.student_price
+    # If price is ``None`` or negative, return a free course.
+    return max(price or 0, 0)
+
+
+@auth.requires_login()
+def payment():
+    # The payment will be made for ``auth.user.course_id``. Get the corresponding course name.
+    course = db(db.courses.id == auth.user.course_id).select(db.courses.course_name).first()
+    assert course.course_name
+    form = StripeForm(
+        pk=settings.STRIPE_PUBLISHABLE_KEY,
+        sk=settings.STRIPE_SECRET_KEY,
+        amount=_course_price(auth.user.course_id),
+        description="Access to the {} textbook from {}".format(course.course_name, settings.title)
+    ).process()
+    if form.accepted:
+        # Save the payment info, then redirect to the index.
+        user_courses_id = db.user_courses.insert(user_id=auth.user.id, course_id=auth.user.course_id)
+        db.payments.insert(user_courses_id=user_courses_id, charge_id=form.response['id'])
+        db.commit()
+        return dict(request=request, course=course, payment_success=True)
+    elif form.errors:
+        return dict(form=form, payment_success=False)
+    # Fix up CSS -- the ``hidden`` attribute hides any error feedback.
+    html = form.xml().replace('"payment-errors error hidden"', '"payment-errors error"')
+    return dict(html=html, payment_success=None)
+
 
 
 @auth.requires_login()
@@ -151,6 +182,13 @@ def index():
         for row in in_db:
             db_check.append(row)
         if not db_check:
+            # The user hasn't been enrolled in this course yet. Check the price for the course.
+            price = _course_price(auth.user.course_id)
+            # If the price is non-zero, then require a payment. Otherwise, ask for a donation.
+            if price > 0:
+                redirect(URL('payment'))
+            else:
+                session.request_donation = True
             db.user_courses.insert(user_id=auth.user.id, course_id=auth.user.course_id)
         try:
             logger.debug("INDEX - checking for progress table")
@@ -169,14 +207,19 @@ def index():
             session.flash = "Your course is not set up to track your progress"
         # todo:  check course.course_name make sure it is valid if not then redirect to a nicer page.
 
-        if session.donate and session.donate != "0":
-            amt = session.donate
-            del session.donate
-            redirect(URL(c='default', f='donate', args=amt))
+        if session.request_donation:
+            del session.request_donation
+            redirect(URL(c='default', f='donate'))
 
-        if session.build:
-            del session.build
+        if session.build_course:
+            del session.build_course
             redirect(URL(c='designer', f='index'))
+
+        # See if we need to do a redirect from LTI.
+        if session.lti_url_next:
+            # This is a one-time redirect.
+            del session.lti_url_next
+            redirect(session.lti_url_next)
 
         # check number of classes, if more than 1, send to course selection, if only 1, send to book
         num_courses = db(db.user_courses.user_id == auth.user.id).count()
@@ -301,16 +344,17 @@ def coursechooser():
 
 @auth.requires_login()
 def removecourse():
-    course_id_query = db(db.courses.course_name == request.args[0]).select(db.courses.id)
-    # todo: properly encode course_names to handle courses with special characters
-    # Check if they're about to remove their currently active course
-    auth_query = db(db.auth_user.id == auth.user.id).select()
-    for row in auth_query:
-        if row.course_name == request.args[0] and course_id_query:
-            session.flash = T("Sorry, you cannot remove your current active course.")
-        else:
-            db((db.user_courses.user_id == auth.user.id) &
-               (db.user_courses.course_id == course_id_query[0].id)).delete()
+    if settings.academy_mode:
+      course_id_query = db(db.courses.course_name == request.args[0]).select(db.courses.id)
+      # todo: properly encode course_names to handle courses with special characters
+      # Check if they're about to remove their currently active course
+      auth_query = db(db.auth_user.id == auth.user.id).select()
+      for row in auth_query:
+          if row.course_name == request.args[0] and course_id_query:
+              session.flash = T("Sorry, you cannot remove your current active course.")
+          else:
+              db((db.user_courses.user_id == auth.user.id) &
+                 (db.user_courses.course_id == course_id_query[0].id)).delete()
 
     redirect('/%s/default/courses' % request.application)
 
@@ -345,47 +389,46 @@ def reportabug():
 
 @auth.requires_login()
 def sendreport():
-    # settings.github_token should be set to a valid Github access token
-    # that has full repo access in models/1.py
-
-    if request.vars['bookerror'] == 'on':
-        basecourse = db(db.courses.course_name == request.vars['coursename']).select().first().base_course
-        if basecourse is None:
-            url = 'https://api.github.com/repos/RunestoneInteractive/%s/issues' % request.vars['coursename']
+    if settings.academy_mode:
+        if request.vars['bookerror'] == 'on':
+            basecourse = db(db.courses.course_name == request.vars['coursename']).select().first().base_course
+            if basecourse is None:
+                url = 'https://api.github.com/repos/RunestoneInteractive/%s/issues' % request.vars['coursename']
+            else:
+                url = 'https://api.github.com/repos/RunestoneInteractive/%s/issues' % basecourse
         else:
-            url = 'https://api.github.com/repos/RunestoneInteractive/%s/issues' % basecourse
-    else:
-        url = 'https://api.github.com/repos/RunestoneInteractive/RunestoneComponents/issues'
-    reqsession = requests.Session()
-    reqsession.auth = ('token', settings.github_token)
-    coursename = request.vars['coursename'] if request.vars['coursename'] else "None Provided"
-    pagename = request.vars['pagename'] if request.vars['pagename'] else "None Provided"
-    details = request.vars['bugdetails'] if request.vars['bugdetails'] else "None Provided"
-    uname = request.vars['username'] if request.vars['username'] else "anonymous"
-    uemail = request.vars['useremail'] if request.vars['useremail'] else "no_email"
-    userinfo = uname + ' ' + uemail
+            url = 'https://api.github.com/repos/RunestoneInteractive/RunestoneComponents/issues'
+        reqsession = requests.Session()
+        reqsession.auth = ('token', settings.github_token)
+        coursename = request.vars['coursename'] if request.vars['coursename'] else "None Provided"
+        pagename = request.vars['pagename'] if request.vars['pagename'] else "None Provided"
+        details = request.vars['bugdetails'] if request.vars['bugdetails'] else "None Provided"
+        uname = request.vars['username'] if request.vars['username'] else "anonymous"
+        uemail = request.vars['useremail'] if request.vars['useremail'] else "no_email"
+        userinfo = uname + ' ' + uemail
 
-    body = 'Error reported in course ' + coursename + ' on page ' + pagename + ' by user ' + userinfo + '\n' + details
-    issue = {'title': request.vars['bugtitle'],
-             'body': body}
-    logger.debug("POSTING ISSUE %s ", issue)
-    r = reqsession.post(url, json.dumps(issue))
-    if r.status_code == 201:
-        session.flash = 'Successfully created Issue "%s"' % request.vars['bugtitle']
-    else:
-        session.flash = 'Could not create Issue "%s"' % request.vars['bugtitle']
-    logger.debug("POST STATUS = %s", r.status_code)
+        body = 'Error reported in course ' + coursename + ' on page ' + pagename + ' by user ' + userinfo + '\n' + details
+        issue = {'title': request.vars['bugtitle'],
+                 'body': body}
+        logger.debug("POSTING ISSUE %s ", issue)
+        r = reqsession.post(url, json.dumps(issue))
+        if r.status_code == 201:
+            session.flash = 'Successfully created Issue "%s"' % request.vars['bugtitle']
+        else:
+            session.flash = 'Could not create Issue "%s"' % request.vars['bugtitle']
+        logger.debug("POST STATUS = %s", r.status_code)
 
-    course_check = 0
-    if auth.user:
-        course_check = db(db.user_courses.user_id == auth.user.id).count()
+        course_check = 0
+        if auth.user:
+            course_check = db(db.user_courses.user_id == auth.user.id).count()
 
-    if course_check == 1 and request.vars['coursename']:
-        redirect('/%s/static/%s/index.html' % (request.application, request.vars['coursename']))
-    elif course_check > 1:
-        redirect('/%s/default/courses' % request.application)
-    else:
-        redirect('/%s/default/' % request.application)
+        if course_check == 1 and request.vars['coursename']:
+            redirect('/%s/static/%s/index.html' % (request.application, request.vars['coursename']))
+        elif course_check > 1:
+            redirect('/%s/default/courses' % request.application)
+        else:
+            redirect('/%s/default/' % request.application)
+    redirect('/%s/default/' % request.application)
 
 
 def terms():
